Orbital element parameters are double

+++ game/orbital_elements.cc
@@ -86,10 +86,10 @@ OrbitalElements orbitalElementsFromStateVector(const glm::dvec3 &r, const glm::d
 void from_json(const nlohmann::json &json, OrbitalElements &elements)
 {
     elements.epoch = JulianDate{JulianClock::duration{json.at("epoch").get<double>()}};
-    elements.semiMajorAxis = json.at("semimajor_axis").get<float>();
-    elements.eccentricity = json.at("eccentricity").get<float>();
-    elements.inclination = glm::radians(json.at("inclination").get<float>());
-    elements.longitudePerihelion = glm::radians(json.at("longitude_perihelion").get<float>());
-    elements.longitudeAscendingNode = glm::radians(json.at("longitude_ascending_node").get<float>());
-    elements.meanAnomalyAtEpoch = glm::radians(json.at("mean_anomaly").get<float>());
+    elements.semiMajorAxis = json.at("semimajor_axis").get<double>();
+    elements.eccentricity = json.at("eccentricity").get<double>();
+    elements.inclination = glm::radians(json.at("inclination").get<double>());
+    elements.longitudePerihelion = glm::radians(json.at("longitude_perihelion").get<double>());
+    elements.longitudeAscendingNode = glm::radians(json.at("longitude_ascending_node").get<double>());
+    elements.meanAnomalyAtEpoch = glm::radians(json.at("mean_anomaly").get<double>());
 }
